fix: Profile view, supporters added and replies image drop fixed and likes replaced with 1 and 0 share button activated

+++ pages/api/like.ts
@@ -18,25 +18,34 @@ export default async function handle(
     return res.status(401).json({ error: "Unauthorized" });
   }
 
+  const post = await Post.findById(id);
+  if (!post) {
+    return res.status(404).json({ error: "Post not found" });
+  }
+
   const existingLike = await Like.findOne({ author: userId, post: id });
 
   if (existingLike) {
-    const deletion = await Like.deleteOne({ author: userId, post: id });
-    if (deletion)
-      await Post.updateOne({ _id: id }, { $inc: { likesCount: -1 } });
-
-    return res.json({ message: "Like removed", likesCount: -1 });
+    if (post.likesCount > 0) {
+      const deletion = await Like.deleteOne({ author: userId, post: id });
+      if (deletion) {
+        await Post.updateOne({ _id: id }, { $inc: { likesCount: -1 } });
+      }
+      return res.json({
+        message: "Like removed",
+        likesCount: post.likesCount - 1,
+      });
+    } else {
+      return res.json({
+        message: "Cannot reduce likes below 0",
+        likesCount: post.likesCount,
+      });
+    }
   } else {
     const creation = await Like.create({ author: userId, post: id });
     if (creation) {
-      await Post.updateOne(
-        { _id: id },
-        {
-          $inc: { likesCount: +1 },
-        }
-      );
+      await Post.updateOne({ _id: id }, { $inc: { likesCount: +1 } });
     }
-
-    return res.json({ message: "Post liked", likesCount: 1 });
+    return res.json({ message: "Post liked", likesCount: post.likesCount + 1 });
   }
 }
